:sparkles: implementing authorization using gates

+++ app/Http/Controllers/QuestionsController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\AskQuestionRequest;
 use App\Models\Question;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -54,7 +55,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
-        return view('questions.edit', compact('question'));
+        if (Gate::allows('update-question', $question)) {
+            return view('questions.edit', compact('question'));
+        }
+        abort(403, 'Access denied');
     }
 
     /**
@@ -74,6 +78,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403, 'Access denied');
+        }
         $question->delete();
         session()->flash('success', 'Your question has been deleted');
